Fixed Integer / Long mixup.

+++ src/main/java/com/github/psygate/bitutils/bitstreams/BitOutputStream.java
@@ -56,7 +56,7 @@ public class BitOutputStream extends OutputStream implements AutoCloseable, Data
 
     public void writeBits(int bits, int amount) throws IOException {
         if (amount < 0 || amount > Integer.SIZE) {
-            throw new IllegalArgumentException("Requested amount exceeds data type size. (" + amount + "/" + Integer.SIZE + ")");
+            throw new IllegalArgumentException("Requested amount exceeds data type size. (" + amount + "/" + Long.SIZE + ")");
         } else if (amount == 0) {
             //Skip
         } else {
